move StreamController to a toplevel type
