+++ gax/src/test/java/com/google/api/gax/rpc/testing/FakeStreamingApi.java
@@ -36,9 +36,9 @@ import com.google.api.gax.rpc.ApiStreamObserver;
 import com.google.api.gax.rpc.BidiStreamingCallable;
 import com.google.api.gax.rpc.ClientStreamingCallable;
 import com.google.api.gax.rpc.ResponseObserver;
-import com.google.api.gax.rpc.ResponseObserver.StreamController;
 import com.google.api.gax.rpc.ServerStreamingCallable;
 import com.google.api.gax.rpc.StatusCode.Code;
+import com.google.api.gax.rpc.StreamController;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Queues;
 import java.util.ArrayList;
